+++ lotto/lotto.py
@@ -1,23 +1,8 @@
 from abc import ABCMeta, abstractmethod
 from typing import List
 
+from lotto.account import Account
 from lotto.types import DateRange
-
-
-class Account(object):
-    def __init__(self, account_id: str, account_password: str) -> None:
-        self.id = account_id
-        self.password = account_password
-
-    def __str__(self) -> str:
-        def mask(string) -> str:
-            return '*' * len(string)
-
-        return f'id={self.id}, ' \
-               f'password={mask(self.password) if self.password else self.password}'
-
-    def __repr__(self) -> str:
-        return f'Account({self})'
 
 
 class Lotto(metaclass=ABCMeta):
