+++ tests/unit/check_lotto_result/test_message.py
