+++ tests/unit/check_lotto_result/test_summary.py
